object: abi: no Routine address type
This reverts commit adafe17814b4cadd5e31f5215e2c3658b12a8424.

+++ internal/isa/arm/macro.go
@@ -182,17 +182,17 @@ func (MacroAssembler) StoreGlobal(f *gen.Func, offset int32, x operand.O) {
 }
 
 func (MacroAssembler) Resume(p *gen.Prog) {
-	padUntil(p, PadWord, int32(abi.TextAddrResume))
+	padUntil(p, PadWord, abi.TextAddrResume)
 	p.Text.PutUint32(in.RET.Rn(RegLink)) // Return from trap handler or import function call.
 }
 
 func (MacroAssembler) Init(p *gen.Prog) {
-	padUntil(p, PadWord, int32(abi.TextAddrStart))
+	padUntil(p, PadWord, abi.TextAddrStart)
 	p.Text.PutUint32(NopWord)
 }
 
 func (MacroAssembler) InitCallEntry(p *gen.Prog) (retAddr int32) {
-	padUntil(p, NopWord, int32(abi.TextAddrEnter))
+	padUntil(p, NopWord, abi.TextAddrEnter)
 
 	// Entry routine
